// src/app/interfaces/NewResponse.ts
export interface NewResponse {
  status?: string;
  totalResults?: number;
  articles: Article[];
}

export interface Article {
  source: Source;
  author?: string;
  title: string;
  description?: string;
  url?: string;
  urlToImage?: string;
  publishedAt?: Date;
  content?: string;
}

export interface Source {
  name: string;
  id: string;
}

export interface ArticlesBtCategoryPage {
  [key: string]: {
    page: number;
    articles: Article[];
  }
}